Using psutil instead of free

+++ scripts/benchmark.py
@@ -29,8 +29,6 @@ def main():
     log_file = Path(args.log)
     log_file.parent.mkdir(parents=True, exist_ok=True)
     tmp_log_file = Path(f"{log_file}.tmp")
-
-    # TODO: this would be done better with a decorator, would avoid the two ifs
     is_benchmarking_pipeline = args.command.split()[0] == "snakemake"
 
     with open(log_file, "w") as log_fh:
@@ -40,30 +38,38 @@ def main():
             "real(s)", "sys(s)", "user(s)", "percent_CPU", "max_RAM(kb)", "FS_inputs", "FS_outputs",
             "elapsed_time_alt(s)"
         ]
+        if is_benchmarking_pipeline:
+            header.append("max_delta_system_RAM(kb)")
         print("\t".join(header), file=log_fh)
 
     time_command = get_time_command()
     benchmark_command = f'{time_command} -o {tmp_log_file} -f "%e\t%S\t%U\t%P\t%M\t%I\t%O"'
 
+    start_time = datetime.datetime.now()
+    main_process = subprocess.Popen(f'{benchmark_command} {args.command}', shell=True)
     if is_benchmarking_pipeline:
         RAM_tmp_log_file = Path(f"{log_file}.RAM.tmp")
-        RAM_benchmarking_process = subprocess.Popen(["python", "scripts/get_RAM_usage.py", str(RAM_tmp_log_file)])
+        RAM_benchmarking_process = subprocess.Popen([sys.executable, "scripts/get_RAM_usage.py", str(RAM_tmp_log_file),
+                                                     str(main_process.pid)])
+    return_code = main_process.wait()
+    if return_code:
+        raise subprocess.CalledProcessError(return_code, main_process.args,
+                                 output=main_process.stdout, stderr=main_process.stderr)
 
-    start_time = datetime.datetime.now()
-    subprocess.check_call(f'{benchmark_command} {args.command}', shell=True)
     end_time = datetime.datetime.now()
     elapsed_seconds = (end_time - start_time).total_seconds()
     with open(tmp_log_file) as log_fh_tmp, open(log_file, "a") as log_fh:
         log_line = log_fh_tmp.readline().strip()
         log_line += f"\t{elapsed_seconds}"
-        print(log_line, file=log_fh)
 
-    if is_benchmarking_pipeline:
-        RAM_benchmarking_process.kill()
-        with open(RAM_tmp_log_file) as log_fh_tmp, open(log_file, "a") as log_fh:
-            print("RAM usage:", file=log_fh)
-            print("\n".join(map(str.strip, log_fh_tmp.readlines())), file=log_fh)
-        RAM_tmp_log_file.unlink()
+        if is_benchmarking_pipeline:
+            RAM_benchmarking_process.kill()
+            with open(RAM_tmp_log_file) as RAM_tmp_log_fh:
+                RAM_usage = RAM_tmp_log_fh.readline().strip()
+            log_line += f"\t{RAM_usage}"
+            RAM_tmp_log_file.unlink()
+
+        print(log_line, file=log_fh)
 
     tmp_log_file.unlink()
